# docs/swagger.yaml
basePath: /
definitions:
  http.APIResponse:
    properties:
      status:
        type: string
    type: object
...
  http.jfrogWebhook:
    properties:
      data:
        properties:
          image_name:
            type: string
          name:
            type: string
          path:
            type: string
          platforms:
            items:
              properties:
                architecture:
                  type: string
                os:
                  type: string
              type: object
            type: array
          repo_key:
            type: string
          sha256:
            type: string
          size:
            type: integer
          tag:
            type: string
        type: object
      domain:
        type: string
      event_type:
        type: string
      jpd_origin:
        type: string
      source:
        type: string
      subscription_key:
        type: string
    type: object
...
  http.quayWebhook:
    properties:
      docker_url:
        type: string
      homepage:
        type: string
      name:
        type: string
      namespace:
        type: string
      repository:
        type: string
      updated_tags:
        items:
          type: string
        type: array
    type: object
  http.registryNotification:
    properties:
      events:
        items:
          properties:
            action:
              type: string
            actor:
              properties:
                name:
                  type: string
              type: object
            id:
              type: string
            request:
              properties:
                addr:
                  type: string
                host:
                  type: string
                id:
                  type: string
                method:
                  type: string
                useragent:
                  type: string
              type: object
            source:
              properties:
                addr:
                  type: string
                instanceID:
                  type: string
              type: object
            target:
              properties:
                digest:
                  type: string
                length:
                  type: integer
                mediaType:
                  type: string
                repository:
                  type: string
                size:
                  type: integer
                tag:
                  type: string
                url:
                  type: string
              type: object
            timestamp:
              type: string
          type: object
        type: array
    type: object
...
  http.trackRequest:
    properties:
      identifier:
        type: string
      provider:
        type: string
      schedule:
        type: string
      trigger:
        type: string
    type: object
  http.trackedImage:
    properties:
      image:
        type: string
      namespace:
        type: string
      policy:
        type: string
      pollSchedule:
        type: string
      provider:
        type: string
      registry:
        type: string
      trigger:
        type: string
    type: object
...
  types.Repository:
    properties:
      digest:
        description: optional digest field
        type: string
      host:
        type: string
      name:
        type: string
      tag:
        type: string
    type: object
...
host: localhost:9300
info:
  contact:
    email: support@keel.sh
    name: Keel Support
    url: https://keel.sh
  description: Automated Kubernetes deployment updates. Keel is a tool for automating
    Kubernetes deployment updates.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://keel.sh/terms/
  title: Keel API
  version: "1.0"
paths:
...
  /v1/tracked:
    get:
      description: Returns a list of all tracked container images with their triggers
        and policies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.trackedImage'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: List tracked images
      tags:
      - tracked
    put:
      consumes:
      - application/json
      description: Updates the trigger type and poll schedule for a tracked resource
      parameters:
      - description: Track settings request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.trackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.APIResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update tracking settings
      tags:
      - tracked
...
  /v1/webhooks/jfrog:
    post:
      consumes:
      - application/json
      description: Receives and processes JFrog Artifactory webhook notifications
        for image push events
      parameters:
      - description: JFrog webhook payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.jfrogWebhook'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Trigger JFrog webhook
      tags:
      - webhooks
  /v1/webhooks/native:
    post:
      consumes:
      - application/json
      description: Triggers a deployment update event directly with repository and
        tag information
      parameters:
      - description: Repository information with name and tag
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.Repository'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Trigger native webhook
      tags:
      - webhooks
  /v1/webhooks/quay:
    post:
      consumes:
      - application/json
      description: Receives and processes Quay.io webhook notifications for image
        push events
      parameters:
      - description: Quay webhook payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.quayWebhook'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Trigger Quay webhook
      tags:
      - webhooks
  /v1/webhooks/registry:
    post:
      consumes:
      - application/json
      description: Receives and processes Docker Registry V2 notification webhooks
        for image push events. Compatible with Docker Registry, GitLab Container Registry,
        and Harbor.
      parameters:
      - description: Registry notification payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.registryNotification'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Trigger Docker Registry notification webhook
      tags:
      - webhooks
...
swagger: "2.0"